src/servers/responseGenerator/requirements.txt
fastapi
uvicorn
pydantic
langchain-core
python-dotenv
langchain
langchain-ollama
ipython